Some quick changes to make all the threads have a joint job pool.
Previously it was the case that some of the threads would finish up
their jobs first, and then be waiting idle while others continued to
work. It also resulted in uneven numbers of samples across the image,
which was problematic for saving out partially rendered images when
using progressive rendering with --display.

TODO: Fix this up! It was a super quick hack.