Phong interpolation, and a long overdue bug fix.
1) I finally added Phong interpolation. It is implemented by computing a vertex
normal for each vertex by taking an unweighted average of the normals of every
triangle that includes the given vertex (where inclusion is evaluated as an
exact floating point match). Finally, barycentric coordinates are computed when
a ray hits a triangle and vertex normals are interpolated as appropriate.

Currently in integrator.cpp this interpolated normal is used for computing
Lambertian and Phong coefficients, but not for computing a scattering direction,
to avoid mini-terminator problems on each triangle. I will have to fix this
later, especially for BSDFs that are sharper (like glossy surfaces).

2) Simple sky color was added to the Scene, that defaults to black.

3) Fixed that horrific bug that made images with multiple lights have weird dark
stripes in them! The issue was fixed by adding the line:

	lambertian_coef = real_max(0.0, lambertian_coef);

Lambertian coefs were ending up negative, and causing the terminator of a light
side from a light source to have a thin strip of negative lighting, causing
weird artifacts. Finally fixed that obnoxious bug.